View,ViewData,ViewZone
Division de ces classes dans les Entity et les Manager
