+++ src/Controller/ViewController.php
@@ -26,6 +26,7 @@ use NextDom\Exceptions\CoreException;
 use NextDom\Helpers\Render;
 use NextDom\Helpers\Status;
 use NextDom\Helpers\Utils;
+use NextDom\Managers\ViewManager;
 
 class ViewController extends BaseController
 {
@@ -51,7 +52,7 @@ class ViewController extends BaseController
     public function get(Render $render, array &$pageContent): string
     {
 
-        $pageContent['viewsList'] = \view::all();
+        $pageContent['viewsList'] = ViewManager::all();
         $pageContent['viewHideList'] = true;
         $pageContent['viewIsAdmin'] = Status::isConnectAdmin();
         $pageContent['viewDefault'] = $_SESSION['user']->getOptions('displayViewByDefault');
@@ -61,14 +62,14 @@ class ViewController extends BaseController
         if (Utils::init('view_id') == '') {
 
             if ($_SESSION['user']->getOptions('defaultDesktopView') != '') {
-                $currentView = \view::byId($_SESSION['user']->getOptions('defaultDesktopView'));
+                $currentView = ViewManager::byId($_SESSION['user']->getOptions('defaultDesktopView'));
             }
 
             if (!is_object($currentView) && is_array($pageContent['viewsList']) && count($pageContent['viewsList']) > 0) {
                 $currentView = $pageContent['viewsList'][0];
             }
         } else {
-            $currentView = \view::byId(init('view_id'));
+            $currentView = ViewManager::byId(init('view_id'));
 
             if (!is_object($currentView)) {
                 throw new \Exception('{{Vue inconnue. Vérifier l\'ID.}}');
